+++ src/Generators/index.js
@@ -390,3 +390,85 @@ generators.listener = {
     return path.join(options.appRoot, options.appDir, options.dirs.listeners, this.getFileName(name)) + '.js'
   }
 }
+
+generators.exceptionHandler = {
+  /**
+   * Returns data for the exception handler template
+   *
+   * @method getData
+   *
+   * @return {Object}
+   */
+  getData () {
+    return {}
+  },
+
+  /**
+   * Returns file name for the exception handler file
+   *
+   * @return {String}
+   */
+  getFileName () {
+    return 'Handler'
+  },
+
+  /**
+   * Returns file path for the exception handler file
+   *
+   * @method getFilePath
+   *
+   * @param  {String}    name
+   * @param  {Object}    options
+   *
+   * @return {String}
+   */
+  getFilePath (name, options) {
+    return path.join(options.appRoot, options.appDir, options.dirs.exceptions, this.getFileName(name)) + '.js'
+  }
+}
+
+generators.seed = {
+  /**
+   * Returns data object for the seed
+   * template file
+   *
+   * @method getData
+   *
+   * @param  {String} name
+   *
+   * @return {Object}
+   */
+  getData (name) {
+    return {
+      name: this.getFileName(name)
+    }
+  },
+
+  /**
+   * Returns the seed file name
+   *
+   * @method getFileName
+   *
+   * @param  {String}    name
+   *
+   * @return {String}
+   */
+  getFileName (name, appPath) {
+    name = name.replace(/seed(er)?/ig, '')
+    return `${pluralize.singular(_.upperFirst(_.camelCase(name)))}Seeder`
+  },
+
+  /**
+   * Returns file path to the model file
+   *
+   * @method getFilePath
+   *
+   * @param  {String}    name
+   * @param  {Object}    options
+   *
+   * @return {String}
+   */
+  getFilePath (name, options) {
+    return path.join(options.appRoot, options.dirs.seeds, this.getFileName(name)) + '.js'
+  }
+}